+++ spec/mailers/alchemy/messages_mailer_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
 module Alchemy
-  describe Messages do
+  describe MessagesMailer do
     let(:message) { Message.new(email: 'jon@doe.com', message: 'Lorem ipsum') }
-    let(:mail) { Messages.contact_form_mail(message, 'admin@page.com', 'contact@page.com', 'Subject') }
+    let(:mail) { MessagesMailer.contact_form_mail(message, 'admin@page.com', 'contact@page.com', 'Subject') }
 
     it "delivers a mail to owner" do
       expect(mail.to).to eq(['admin@page.com'])